Extract current item
