+++ Sources/Player/Player+ItemUpdate.swift
@@ -12,11 +12,11 @@ extension Player {
     struct ItemUpdate {
         let items: Deque<PlayerItem>
         let currentItem: AVPlayerItem?
-        
+
         func currentIndex() -> Int? {
             items.firstIndex { $0.matches(currentItem) }
         }
-        
+
         func streamTypePublisher() -> AnyPublisher<StreamType, Never> {
             guard let currentItem else { return Just(.unknown).eraseToAnyPublisher() }
             return currentItem.streamTypePublisher().eraseToAnyPublisher()
@@ -26,7 +26,7 @@ extension Player {
     func itemUpdatePublisher() -> AnyPublisher<ItemUpdate, Never> {
         Publishers.CombineLatest($storedItems, $currentItem)
             .map { items, currentItem in
-                let playerItem = Self.smoothPlayerItem(for: currentItem, in: items)
+                let playerItem = currentItem.smoothPlayerItem(in: items)
                 return ItemUpdate(items: items, currentItem: playerItem)
             }
             .eraseToAnyPublisher()
